[flutter_local_notifications] DST, function typo, header and newline consistency in readme (#1571)
* DST and function typo

My understanding is that the best way of referring to DST is not "daylight savingS" but "daylight saving", and I put in "time" for completeness for a readme. https://en.wikipedia.org/wiki/Daylight_saving_time

* Linearly changing header levels; consistent newlines

Some headers were level 4 and 5 after a level 2 header, so I bumped them up. I removed extra newlines and added missing ones since most of the doc uses the style of 2 newlines after a header.